requirements.txt
aiogram
python-dotenv
g4f
apscheduler